+++ source/serializables/command.js
@@ -1,0 +1,6 @@
+
+Space.Struct.extend('Space.messaging.Command', {
+
+  mixin: [Space.messaging.Ejsonable]
+
+});